Git Init + Working Script
The Script Works as expected in this stage and passes all 6 tests. I'm
open to adding more tests if other submit them. Next to do is to write
documentation and see about making the type and format section an org
table rather than a python dicionary. That way it would be a matter of
changing the info in a table. As is though it should be sutible for most.